Makefile: use absolute path for openocd
OpenOCD assumes all files are relative to the directory it is running
in, not from where gdb is running from. Use absolute paths when
programming so OpenOCD doesn't need to be running from the same
directory.

This enables it to be used when OpenOCD is running from a systemd unit
file.