*: allow for arbitarily long adc captures
Instead of capturing data into a buffer then processing it, process the
data as it comes. Since we're just calculating the DFT at one frequency
we can just accumulate the products of f(x) and e^(iwf) as we get data.

This has two main advantages:
 1) We can capture more data than would fit in RAM, improving precision.
 2) We free up the 64k of ram used to the ADC buffer. This could be used
    to store sin/cos tables so they only need to be calculated once.

Since the ADC is running at 60 cycles/sample we need to process the data
at least that fast. Use a second lookup table for cos instead of using
an offset into the sine. Since ADC data arrives 4x slower than DAC data,
we can also use 4x smaller tables in the future, which should allow 32
tables to neatly fit in 64k of ram.